Ditch h2-test-support crate for support mod
The h2-test-support caused the unstable flag to always be enabled.
