+++ tests/flow_control.rs
@@ -1,6 +1,6 @@
 #[macro_use]
-extern crate h2_test_support;
-use h2_test_support::prelude::*;
+pub mod support;
+use support::prelude::*;
 
 // In this case, the stream & connection both have capacity, but capacity is not
 // explicitly requested.
