+++ tests/push_promise.rs
@@ -1,5 +1,5 @@
-extern crate h2_test_support;
-use h2_test_support::prelude::*;
+pub mod support;
+use support::prelude::*;
 
 #[test]
 fn recv_push_works() {
