--- tests/support/README.md
@@ -1,3 +1,0 @@
-**This is not a real crate**
-
-You should not use this crate, it only exists to work around cargo limitations.
