+++ tests/support/raw.rs
@@ -7,19 +7,19 @@ macro_rules! raw_codec {
             $fn:ident => [$($chunk:expr,)+];
         )*
     ) => {{
-        let mut b = $crate::mock_io::Builder::new();
+        let mut b = $crate::support::mock_io::Builder::new();
 
         $({
             let mut chunk = vec![];
 
             $(
-                $crate::raw::Chunk::push(&$chunk, &mut chunk);
+                $crate::support::raw::Chunk::push(&$chunk, &mut chunk);
             )+
 
             b.$fn(&chunk[..]);
         })*
 
-        $crate::Codec::new(b.build())
+        $crate::support::Codec::new(b.build())
     }}
 }
