+++ tests/trailers.rs
@@ -1,8 +1,8 @@
 #[macro_use]
 extern crate log;
 
-extern crate h2_test_support;
-use h2_test_support::prelude::*;
+pub mod support;
+use support::prelude::*;
 
 #[test]
 fn recv_trailers_only() {
